fix: handleError returns Error objects instead of strings
The errors returned by the API contain the status code and the JSON string body which makes it difficult to parse the message or the status code correctly. I’ve changed this so that the error is an instance of the Error object which contains the "error" from the response as message and the whole response in "err.response". This way libraries using the api can handle the error and also react on the response.